refactor : User 클래스를 DB 테이블에 매핑
